backend/requirements.txt
# [Core Web Framework]
# FastAPI for the API, Uvicorn for the server.
fastapi==0.111.1
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# [Settings and Data Validation]
# Pydantic for data models, pydantic-settings for environment management.
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1

# [Database & ORM]
# SQLModel for the SQLite ORM, SQLAlchemy is its core dependency.
sqlmodel==0.0.21
sqlalchemy==2.0.31

# [Vector Store Client]
# ChromaDB client, updated to the latest major version.
# IMPORTANT: This requires updating the ChromaDB image in docker-compose.yml.
chromadb==1.0.17
# [Document Parsers]
# Libraries for extracting text from various file formats.
pypdf==4.3.1
python-docx==1.1.2
beautifulsoup4==4.12.3
markdown==3.6
pdfplumber==0.11.7
pdfminer.six==20221105

# [OCR Engine]
# Optional: PaddleOCR for image-to-text conversion.
# IMPORTANT: paddlepaddle must be installed separately based on your system (CPU/GPU).
# See https://www.paddlepaddle.org.cn/install/quick for instructions.
paddleocr==2.8.0
# Example CPU install:
# paddlepaddle==2.6.1

# [RAG & Embeddings]
# SentenceTransformers for embeddings, Rank-BM25 for keyword ranking.
rank-bm25==0.2.2
click==8.2.1
h11==0.16.0
starlette==0.47.2
typing-extensions==4.14.1
pydantic-core==2.20.1
overrides==7.7.0
tenacity==9.1.2
httpx==0.28.1
onnxruntime==1.22.1
tokenizers==0.21.4
lxml==6.0.0
torch==2.8.0
transformers==4.55.2
scikit-learn==1.7.1
pillow==11.3.0
numpy==2.3.2
anyio==4.10.0
pybase64==1.4.2
pyyaml==6.0.2
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
cryptography==45.0.6
tqdm==4.67.1
huggingface-hub==0.34.4
annotated-types==0.7.0
sniffio==1.3.1
jsonschema
attrs==25.3.0
referencing==0.36.2
jsonschema-specifications==2025.4.1
idna==3.10
importlib-metadata==8.7.0
grpcio==1.74.0
google==3.0.0
protobuf==6.32.0
googleapis-common-protos==1.70.0
charset-normalizer==3.4.3
cffi==1.17.1
requests==2.32.4
filelock==3.19.1
packaging==25.0
regex==2025.7.34
safetensors==0.6.2
rpds-py==0.27.0
urllib3==2.5.0
certifi==2025.8.3
sentence-transformers
posthog==6.5.0
orjson==3.11.2
httpcore==1.0.9

accelerate==0.32.1
bitsandbytes==0.43.1
einops==0.8.0
ctransformers